# src/docs/swagger.yaml
openapi: 3.0.0
info:
    title: House Sale API Documentation
    version: 1.0.0
servers:
    -   url: http://localhost:3000/v1/api
    -   url: https://housesale.tldev.id.vn/v1/api
...
paths:
...
    /user/register-user:
        post:
            tags:
                - User
            summary: Create new user
            description: Create new user
            operationId: registerUser
            requestBody:
                description: User email and password
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserBody'
                required: true
            responses:
                '201':
                    description: Registration success for new user!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example: Registration success for new seller!
                                    metaData:
                                        type: object
                                        properties:
                                            newUser:
                                                $ref: '#/components/schemas/User'
                                            tokens:
                                                $ref: '#/components/schemas/Tokens'
                '400':
                    description: DuplicateEmail
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: User already exists! Please register with another email.
    /user/register-seller:
        post:
            tags:
                - User
            summary: Create new seller
            description: Create new seller, send verification email to seller's email
            operationId: registerSeller
            requestBody:
                description: Seller information
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NewSeller'
                required: true
            responses:
                '201':
                    description: Registration success for new seller! Please check your email to verify your account!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example: Registration success for new seller!
                                    metaData:
                                        type: object
                                        properties:
                                            newSeller:
                                                $ref: '#/components/schemas/Seller'
                                            tokens:
                                                $ref: '#/components/schemas/Tokens'
                '400':
                    description: DuplicateEmail
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: User already exists! Please register with another email.
...
    /user/refreshTokens:
        post:
            tags:
                - User
            summary: Refresh Tokens
            description: Refresh Tokens
            operationId: refreshTokens
            requestBody:
                description: refreshToken
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                              refreshToken:
                                type: string
                                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjIsImlhdCI6MTcwNzQ0Mzc0MywiZXhwIjoxNzA4MDQ4NTQzLCJ0eXBlIjoicmVmcmVzaCJ9.xbR_ZLbKphz3161exK3HpoWPmh8GH6w3Gu0USDv8nko
                            required:
                              - "refreshToken"
                required: true
            responses:
                '201':
                    description: Create new accessToken success!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 201
                                    message:
                                        type: string
                                        example: Create new accessToken success!
                                    metaData:
                                        type: object
                                        properties:
                                            tokens:
                                                $ref: '#/components/schemas/Tokens'
                '400':
                    description: Failed creating new tokens
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: Failed creating new tokens
                '404':
                    description: Tokens or user not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 401
                                            message:
                                                type: string
                                                example: Tokens not found
    /user/verify-email/{userId}/{code}:
        get:
            tags:
                - User
            summary: Verify seller email when register seller account
            description: System send a verify link to seller's email, seller must verify email to access system
            operationId: verifyEmail
            parameters:
                -   name: userId
                    in: path
                    description: ID of user
                    required: true
                    schema:
                        type: number
                        example: 44
                -   name: code
                    in: path
                    description: code to verify email
                    required: true
                    schema:
                        type: string
                        format: uuid
                        example: e916faf0-3123-4fcc-8ec3-9a2c2bb4b38c44
            responses:
                '200':
                    description: Your email had been verified successfully!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example: Your email had been verified successfully!
                                    metaData:
                                        type: object
                                        example: {}

                '400':
                    description: Your email already verified! or Verify email failed!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: Your email already verified!
                '401':
                    description: Incorrect email verification code!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 401
                                            message:
                                                type: string
                                                example: Incorrect email verification code!
...
    /user/change-password:
        post:
            tags:
                - User
            summary: Change user password
            description: Change user password
            operationId: changePassword
            requestBody:
                description: current password and new password
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                currentPassword:
                                    type: string
                                    example: T12345678
                                newPassword:
                                    type: string
                                    example: T123456789
                            required:
                              - "currentPassword"
                              - "newPassword"
                required: true
            responses:
                '200':
                    description: 'Your password has been changed successfully!'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example: 'Your password has been changed successfully!'
                                    metaData:
                                        type: object
                                        example: {}
                '400':
                    description: Failed update password
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: Failed update password
                '401':
                    description: Unauthorized or Incorrect current password
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 401
                                            message:
                                                type: string
                                                example:  Incorrect current password
                '404':
                    description: Not found user
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 404
                                            message:
                                                type: string
                                                example: Not found user
            security:
                - bearerAuth: []
    /user/profile:
        get:
            tags:
            -   User
            summary: Get user profile  by ID
            description: Returns a user profile
            operationId: getProfile
            responses:
                '200':
                    description: Get your profile successfully!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example:  Get your profile successfully!
                                    metaData:
                                        $ref: '#/components/schemas/UserProfile'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 401
                                            message:
                                                type: string
                                                example:  Please Authenticate!
                '404':
                    description: User profile not found or User not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: User profile not found!
            security:
              - bearerAuth: []
        patch:
            tags:
            -   User
            summary: Update user profile  by ID
            description: Update a user profile
            operationId: updateProfile
            requestBody:
                content:
                  application/json:
                    schema:
                        $ref: '#/components/schemas/UpdateUserProfile'
            responses:
                '200':
                    description: Your profile had been changed successfully!
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example:  Your profile had been changed successfully!
                                    metaData:
                                        type: object
                                        example: {}
                '400':
                    description: Failed update phone number or new phone number cannot be same as your current phone number or Invalid location provided.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: New phone number cannot be same as your current phone number!
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 401
                                            message:
                                                type: string
                                                example:  Please Authenticate!
                '404':
                    description: User profile not found or User not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: User profile not found!
            security:
              - bearerAuth: []
    /user/update-avatar:
        post:
            tags:
                - User
            summary: Update user avatar
            description: Update user avatar
            operationId: updateAvatar
            requestBody:
                description: new avatar url
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                imageUrl:
                                    type: string
                                    example: https://res.cloudinary.com/housesale/image/upload/v1706774835/avatar/32/avatar.png
                            required:
                              - "imageUrl"
                required: true
            responses:
                '200':
                    description: 'Your avatar had been changed successfully!'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    status:
                                        type: number
                                        example: 200
                                    message:
                                        type: string
                                        example: Your avatar had been changed successfully!
                                    metaData:
                                        type: object
                                        example: {}
                '400':
                    description: Update your avatar failed
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 400
                                            message:
                                                type: string
                                                example: Update your avatar failed
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 401
                                            message:
                                                type: string
                                                example:  Please Authenticate!
                '404':
                    description: User not found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: object
                                        properties:
                                            status:
                                                type: number
                                                example: 404
                                            message:
                                                type: string
                                                example: User not found
            security:
                - bearerAuth: []
...
components:
    schemas:
...
        User:
            type: object
            properties:
                userId:
                    type: number
                    example: 15
                roleId:
                    type: number
                    example: 1
                email:
                    type: string
                    example: tldevuser@gmail.com
                status:
                    type: boolean
                    example: true
        Tokens:
            type: object
            properties:
                accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjE1LCJpYXQiOjE3MDcyOTE4MzYsImV4cCI6MTcwNzI5NTQzNiwidHlwZSI6ImFjY2VzcyJ9.T8u9lbLeeBXD4peVgar6YX6ZTWvMOeIjOAF_nXcIu9s
                refreshToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjE1LCJpYXQiOjE3MDcyOTE4MzYsImV4cCI6MTcwNzg5NjYzNiwidHlwZSI6InJlZnJlc2gifQ.iHT53WUnZjRbjfhc9Qrb-BUYpWCabsKmZMeRiA7wpQQ
        Seller:
            type: object
            properties:
                status:
                    type: boolean
                    example: true
                userId:
                    type: number
                    example: 21
                email:
                    type: string
                    example: tldevseller@gmail.com
                fullName:
                    type: string
                    example: Test Seller
                phone:
                    type: string
                    example: "0888888776"
                provinceCode:
                    type: string
                    example: 48
                districtCode:
                    type: string
                    example: 492
                wardCode:
                    type: string
                    example: 20236
                street:
                    type: string
                    example: Bach Dang
                address:
                    type: string
                    example: 107
                roleId:
                    type: number
                    example: 2
                updatedAt:
                    type: string
                    format: date-time
                    example: "2024-02-07T08:08:44.032Z"
                createdAt:
                    type: string
                    format: date-time
                    example: "2024-02-07T08:08:44.032Z"
        NewSeller:
            type: object
            properties:
                email:
                    type: string
                    example: tldevseller@gmail.com
                password:
                    type: string
                    example: T12345678
                fullName:
                    type: string
                    example: Test Seller
                phone:
                    type: string
                    example: "0888888776"
                provinceCode:
                    type: string
                    example: "48"
                districtCode:
                    type: string
                    example: "492"
                wardCode:
                    type: string
                    example: "20236"
                street:
                    type: string
                    example: Bach Dang
                address:
                    type: string
                    example: "107"
            required:
              - "email"
              - "password"
              - "fullName"
              - "phone"
              - "provinceCode"
              - "districtCode"
              - "wardCode"
              - "street"
        UserBody:
            type: object
            properties:
                email:
                    type: string
                    example: tldevuser@gmail.com
                password:
                    type: string
                    example: T12345678
            required:
              - "email"
              - "password"
        UserProfile:
            type: object
            properties:
                userId:
                    type: number
                    example: 2
                fullName:
                    type: string
                    example: "Le Nhat Trinh"
                email:
                    type: string
                    example: "mail.tldev@gmail.com"
                phone:
                    type: string
                    example: "0899619252"
                avatar:
                    type: string
                    example: null
                street:
                    type: string
                    example: "Bach Dang"
                address:
                    type: string
                    example: "107"
                status:
                    type: boolean
                    example: true
                createdAt:
                    type: string
                    format: date-time
                    example: "2024-01-31T08:00:12.000Z"
                updatedAt:
                    type: string
                    format: date-time
                    example: "2024-02-09T05:10:01.000Z"
                role:
                    type: object
                    properties:
                        roleId:
                            type: number
                            example: 2
                        roleName:
                            type: string
                            example: "Seller"
                ward:
                    type: object
                    properties:
                        wardCode:
                            type: string
                            example: "20236"
                        fullNameEn:
                            type: string
                            example: "Hai Chau I Ward"
                province:
                    type: object
                    properties:
                        provinceCode:
                            type: string
                            example: "48"
                        fullNameEn:
                            type: string
                            example: "Da Nang City"
                district:
                    type: object
                    properties:
                        districtCode:
                            type: string
                            example: "492"
                        fullNameEn:
                            type: string
                            example: "Hai Chau District"
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT